--- .gitignore
@@ -1,4 +1,0 @@
-.buildpath
-.project
-.settings
-/.DS_Store
