+++ src/Opensearchserver/OssExamples.php
@@ -20,14 +20,13 @@
 *  along with OpenSearchServer PHP Client.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-
+ 
 /**
  * @file
  * Examples using the OpenSearchServer PHP API
  */
 
-require_once(dirname(__FILE__).'/oss_api.class.php');
-require_once(dirname(__FILE__).'/oss_results.class.php');
+namespace Opensearchserver;
 
 // Retrieve connection information
 $oss_url = getenv('OSS_PHP_URL');
@@ -51,15 +50,9 @@ $oss_result = new OssResults($xmlResult);
 $doc_found_number = $oss_result->getResultFound() - $oss_result->getResultCollapsedCount();
 print 'Documents found: '.$doc_found_number."\n";
 
+// Print the title of the documents
 for ($i = 0; $i < $doc_found_number; $i++) {
-  // Print the title of the documents
   $title = $oss_result->getField($i, 'title');
   print '#'.$i.' '.$title."\n";
-  
-  //Example for a multi-valued field : 
-  $categories = $oss_result->getField($i, 'category', false, false, null, true);
-  $categoriesText = (is_array($categories)) ? implode(', ', $categories) : $categories;
-  print 'Category: '.$categoriesText."\n";
- 
 }
