+++ src/Opensearchserver/OssSchema.php
@@ -25,7 +25,7 @@
  * Class to access OpenSearchServer API
  */
 
-require_once(dirname(__FILE__).'/oss_abstract.class.php');
+namespace Opensearchserver;
 
 class OssSchema extends OssAbstract {
